refactor(docs-panel): extract DocsPanelContentArea component
Extract the 5+ branch content switch (~340 LoC) into a focused
component. Renders one of:
  1. FullScreenModeNotice (when full-screen owns the session)
  2. ContextPanel (recommendations tab)
  3. Dev Tools (SelectorDebugPanel, lazy)
  4. Editor (BlockEditor, lazy)
  5. Loading skeleton (with milestone bar still visible for journey
     tabs reloading between milestones)
  6. Error state with retry
  7. ContentRenderer + content meta + milestone toolbar + footer

The lazy() imports for SelectorDebugPanel and BlockEditor move into the
new component file using the same relative module specifiers — webpack
chunk resolution remains stable (pre-mortem H8 verification step).

testIds moved from docs-panel.tsx -> components/DocsPanelContentArea.tsx:
  - testIds.docsPanel.content
  - testIds.devTools.previewBanner
  - testIds.devTools.previewModeIndicator
  - testIds.devTools.returnToEditorButton

SOURCE_CONTRACT in docs-panel.contract.test.tsx is updated in the same
commit to reflect the new owner.

Imports dropped from docs-panel.tsx as a result of the move:
  - SkeletonLoader, ContentRenderer, AlignmentPendingContext
  - LoadingIndicator, ErrorDisplay, AlignmentPrompt, FullScreenModeNotice
  - LearningJourneyMilestoneToolbar, PanelModeActionButtons,
    DocsPanelHeaderMenu (now only used inside the content area)
  - Icon, IconButton, Button (from @grafana/ui — no remaining JSX uses)
  - lazy SelectorDebugPanel/BlockEditor declarations (moved with the
    component, NOT duplicated)
  - getMilestoneSlug, markMilestoneDone, isLastMilestone — now used
    inside the component or the auto-complete hook
  - pickGrafanaDocsOpenAction — now used inside the component

Phase: 1 — Step 1.12
Pattern: E — interface-first component extraction (largest of Phase 1)
Risk: low — JSX moved verbatim, all behavior preserved, lazy() module
  specifiers unchanged to preserve webpack chunk identity (H8)
Tripwire: docs-panel.contract.test.tsx (testId + window-global +
  scroll-target exhaustiveness) passes after SOURCE_CONTRACT update.
  All existing tests for the extracted components
  (LoadingIndicator/ErrorDisplay/AlignmentPrompt/etc) continue to pass.
Checkpoint: prettier + typecheck + lint + test:ci src/components/docs-panel
  -> 36 suites / 327 passed / 6 skipped
  (delta +1 contract entry vs Step 1.11; no regressions)
  docs-panel.tsx LoC: 2681 (baseline) -> 1824 (-857, -32%)

See plan: ~/.claude/plans/please-make-a-plan-elegant-lemon.md
Guidelines: https://github.com/grafana/grafana-pathfinder-app/wiki/High-Risk-Refactor-Guidelines
